+++ NEWS.md
@@ -5,6 +5,7 @@
 
 * `pivot()`
 
+  * `pivot_long()` gracefully handles duplicated column names (#472)
   * `pivot_wide()` can aggregate (#474)
   * `pivot_wide()` can select keys (#572) 
   * Better control of expansion of factor levels not present in the data (#193)